kzalloc checks
SVN revision 6080

+++ src/kernel/linux-2.6/pvfs2-bufmap.c
@@ -63,7 +63,7 @@ static int initialize_bufmap_descriptors(int ndescs)
     if (desc_array == NULL)
     {
         gossip_err("pvfs2: could not allocate %d bytes\n",
-                ndescs * sizeof(struct pvfs_bufmap_desc));
+		(int) (ndescs * sizeof(struct pvfs_bufmap_desc)));
         goto out1;
     }
     err = 0;
